Fix: api 제한 수정

See merge request s04-webmobile2-sub3/s04p13c108!311

+++ backend/src/main/java/com/ssafy/doit/controller/ProductController.java
@@ -270,6 +270,10 @@ public class ProductController {
     public Object cancel(@RequestParam Long roomPk){
         ResponseBasic result = null;
         try {
+            User currentUser = userRepository.findById(userService.currentUser()).get();
+            if (!chatRoomJoinRepository.findChatRoomJoinByUser_IdAndChatRoom_Id(currentUser.getId(), roomPk).isPresent())
+                throw new Exception("유저 불일치");
+
             ChatRoom chatRoom = chatRoomRepository.findById(roomPk).get();
             Sale sale = saleRepository.findByChatRoom_Id(chatRoom.getId()).get();
             saleRepository.delete(sale);
